+++ src/com/shopbilling/services/ProductSaleAnalysisServices.java
@@ -39,7 +39,7 @@ public class ProductSaleAnalysisServices {
 				while (rs.next()) {
 					product = new ProductAnalysis();
 					product.setProductCode(rs.getInt("ITEM_NUMBER"));
-					product.setTotalQty(rs.getInt("TOTAL_QTY"));
+					product.setTotalQty(rs.getDouble("TOTAL_QTY"));
 					
 					productAnalysisList.add(product);
 				}
@@ -105,8 +105,8 @@ public class ProductSaleAnalysisServices {
 					product.setProductCode(rs.getInt("ITEM_NUMBER"));
 					product.setProductName(rs.getString("PRODUCT_NAME"));
 					product.setProductMRP(rs.getDouble("ITEM_MRP"));
-					product.setTotalQty(rs.getInt("TOTAL_QTY"));
-					product.setCurrentQty(rs.getInt("QUANTITY"));
+					product.setTotalQty(rs.getDouble("TOTAL_QTY"));
+					product.setCurrentQty(rs.getDouble("QUANTITY"));
 					
 					productAnalysisList.add(product);
 				}
